navigation ready in sigup and home screen
