+++ RSVflights/App.js
@@ -1,40 +1,27 @@
-import React, {useState, useEffect} from 'react';
-import {StyleSheet, SafeAreaView, StatusBar, Text, View} from 'react-native';
-import Auth from './src/components/Auth';
-import firebase from './src/utils/firebase';
-import 'firebase/auth';
-export default function App() {
-  const [user, setUser] = useState(undefined);
+import React from 'react';
+// import { Text, View } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack'
+import 'react-native-gesture-handler';
+import SingUp from './src/screens/SingUp';
+import Home from './src/screens/Home';
+import LogIn from './src/screens/LoginForm';
+import Booking from './src/screens/Booking'
 
-  useEffect(() => {
-    firebase.auth().onAuthStateChanged((response) => {
-      setUser(response);
-    });
-  }, []);
+const Stack = createStackNavigator();
 
-  if (user === undefined) {
-    return null;
-  }
+const App = () => {
   return (
-    <>
-      <StatusBar barStyle="light-content" />
-      <SafeAreaView style={styles.background}>
-        {user ? (
-          <View>
-            <Text>Logeado</Text>
-            <Text onPress={() => firebase.auth().signOut()}>Cerrar Sesión</Text>
-          </View>
-        ) : (
-          <Auth />
-        )}
-      </SafeAreaView>
-    </>
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name ="LogIn" component ={LogIn} options= {{headerShown: false}}  />
+        <Stack.Screen name ="SingUp" component ={SingUp} options= {{headerShown: false}} />
+        <Stack.Screen name ="Home" component ={Home} options= {{headerShown: false}}   />
+        <Stack.Screen name ="Booking" component ={Booking} options= {{headerTitle: ''}}  />
+        
+      </Stack.Navigator>
+    </NavigationContainer>
   );
-}
+};
 
-const styles = StyleSheet.create({
-  background: {
-    backgroundColor: '#fff',
-    height: '100%',
-  },
-});
+export default App;
