+++ RSVflights/redux/app-redux.js
@@ -1,0 +1,3 @@
+import { createStore, applyMiddleware } from 'redux';
+import thunkMiddleware from 'redux-thunk';
+
